+++ cf-eventhub/src/index.ts
@@ -9,11 +9,13 @@ export type {
   ListedPayload,
 } from "./core/store";
 export type { EventPayload, JSONObject } from "./core/type";
-export { configureDelivery, EventHub } from "./eventhub";
+export { configureDelivery, configureEviction, EventHub } from "./eventhub";
 export type {
   DeliveryConfig,
   EjectOptions,
   ListEjectedOptions,
   ListOrder,
   ListOptions,
+  EvictionAction,
+  EvictionConfig,
 } from "./eventhub";